Changed the parse with a method with a return type of decimal from int32.parse to decimal.parse

+++ Lemonade Stand/Lemonade Stand.cs
@@ -13,7 +13,7 @@ namespace ConsoleApp15
         
         public decimal Inventory()
         {
-            return Int32.Parse(strInventory);
+            return decimal.Parse(strInventory);
         }
 
         public decimal CostOfProduction()
